requirements.txt
playwright==1.58.0
pdfplumber>=0.9.0
python-docx>=0.8.11
asyncio
flask
flask-cors
requests
